requirements.txt
beautifulsoup4==4.9.0
click==7.1.1
Flask==1.1.1
google==2.0.3
grpcio==1.28.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.11.3
six==1.14.0
soupsieve==2.0
Werkzeug==1.0.1
